// backend/src/models/types/auth-type.ts
export default interface AuthType {

	username: string
	email: string
	password: string
	ispartner: string
	userhash: string
	activated: string
	isbanned: string
	banreason: string
	newpasswordkey: string
	newpasswordkeyrequested: string
	activationemailkey: string
	emailkeyrequested: string
	qrcode: string
	lastip: string
	useragent: string
	lastlogin: string
	created_at: string
	modified_at: string
	accept_terms: string
	accept_terms_date: string
	isactive: string

}